% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{model}
\alias{model}
\title{Hourly simulated Precipitation data}
\format{
A data frame with 288 rows and 2 columns:
\describe{
  \item{DateTime}{Date Time in YYYY-MM-DD hh:mm:ss at 1-hour intervals.}
  \item{PRECIP}{Simulated precipitation data.}
  }
}
\source{
https://doi.org/10.1038/s41597-022-01304-7
}
\usage{
model
}
\description{
Precipitation data for a station recorded at 1-hour intervals for 2030.
}
\keyword{datasets}
