% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs15.R
\docType{data}
\name{obs15}
\alias{obs15}
\title{15-min Observed Precipitation data}
\format{
A data frame with 10,086 rows and 2 columns:
\describe{
  \item{DateTime}{Date Time in YYYY-MM-DD hh:mm:ss at 15-min intervals.}
  \item{PRECIP}{Observed precipitation data.}
}
}
\source{
https://doi.org/10.1038/s41597-022-01304-7
}
\usage{
obs15
}
\description{
Precipitation data for a station recorded at 15-min intervals from 1981-2000.
}
\keyword{datasets}
