% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_30.R
\docType{data}
\name{obs_30}
\alias{obs_30}
\title{30-min Observed Precipitation data}
\format{
A data frame with 8,881 rows and 2 columns:
\describe{
  \item{DateTime}{DateTime in YYYY-MM-DD hh:mm:ss at 30-min intervals.}
  \item{PRECIP}{Observed precipitation data.}
}
}
\source{
https://doi.org/10.1038/s41597-022-01304-7
}
\usage{
obs_30
}
\description{
Precipitation data for a station recorded at 30-min intervals from 1981-2000.
}
\keyword{datasets}
