% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed.R
\docType{data}
\name{observed}
\alias{observed}
\title{15-min Observed Precipitation data}
\format{
A data frame with 10,086 rows and 2 columns:
\describe{
  \item{DT}{Date Time in YYYY-MM-DD hh:mm:ss}
  \item{PRECIP}{Observed precipitation data recorded at 15-min intervals}
}
}
\source{
https://doi.org/10.1038/s41597-022-01304-7
}
\usage{
observed
}
\description{
Precipitation data for a station recorded at 15-min intervals from 1981-2000.
}
\keyword{datasets}
