% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precip_disagg.R
\name{Preci_diasg}
\alias{Preci_diasg}
\title{Temporal downscaling of precipitation.}
\usage{
Preci_diasg(obs, mod)
}
\arguments{
\item{obs}{A \code{dataframe} of dimension 2 used to denote the observed precipitation
that recorded in 15-min.
This can be either continuous or discontinuous time series data frame.}

\item{mod}{A \code{dataframe} of dimension 2 used to denote the observed precipitation
that recorded in 1-h.
This can be either continuous or discontinuous time series data frame.}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{I}{Estimated value of the integral}
     \item{var}{Estimated variance of the estimator}
}
}
\description{
Preci_diasg () generates a higher resolution temporal scale (15-min) from 1-h
 precipitation using a modified stochastic disaggregation method.
}
\examples{
Preci_diasg(obs=observed,mod=model)
}
\author{
Bijoychandra Takhellambam, Puneet Srivastava,
Jasmeet Lamba, Ryan P. McGehee, Hemendra Kumar, & Di Tian.
}
