% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecipTDS15.R
\name{PrecipTDS15}
\alias{PrecipTDS15}
\title{Temporal downscaling of precipitation from 1-hour to 15-min.}
\usage{
PrecipTDS15(obs, mod)
}
\arguments{
\item{obs}{A \code{dataframe/tibble} of dimension 2 used to denote the observed precipitation
that recorded in 15-min.
This can be either continuous or discontinuous time series data frame.}

\item{mod}{A \code{dataframe/tibble} of dimension 2 used to denote the simulated precipitation
that recorded in 1-h.
This can be either continuous or discontinuous time series data frame.}
}
\value{
PrecipTDS15 returns a continuous time series of precipitation.A \code{tibble} containing the following attributes:
\describe{
     \item{Datetime}{The generated 15-min continous date and time.}
     \item{Preci}{Temporally downscaled precipitation  at 15-min interval.}
}
}
\description{
PrecipTDS15 () generates a higher resolution temporal scale (15-min) from 1-h
 precipitation using a modified stochastic disaggregation method.
}
\examples{
PrecipTDS15(obs=obs_15,mod=model)
}
\author{
Bijoychandra Takhellambam, Puneet Srivastava,
Jasmeet Lamba, Hemendra Kumar, & Roberto Molinari.
}
