% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecipTDS30.R
\name{PrecipTDS30}
\alias{PrecipTDS30}
\title{Temporal downscaling of precipitation from 1-hour to 30-min.}
\usage{
PrecipTDS30(obs, mod)
}
\arguments{
\item{obs}{A \code{dataframe/tibble} of dimension 2 used to denote the observed precipitation
that recorded in 30-min.
This can be either continuous or discontinuous time series data frame.}

\item{mod}{A \code{dataframe/tibble} of dimension 2 used to denote the simulated precipitation
that recorded in 1-h.
This can be either continuous or discontinuous time series data frame.}
}
\value{
PrecipTDS30 returns a continuous time series of precipitation.A \code{tibble} containing the following attributes:
\describe{
     \item{Datetime}{The generated 30-min continous date and time.}
     \item{Preci}{Temporally downscaled precipitation  at 30-min interval.}
}
}
\description{
PrecipTDS30 () generates a higher resolution temporal scale (30-min) from 1-h
 precipitation using a modified stochastic disaggregation method.
}
\examples{
PrecipTDS30(obs=obs30,mod=model)
}
\author{
Bijoychandra Takhellambam, Puneet Srivastava,
Jasmeet Lamba, Hemendra Kumar, & Roberto Molinari.
}
